refactor: refactoring the process config structure and tests
