+++ packages/config/src/configProcessor/helpers/behaviors.ts
