+++ packages/config/src/configProcessor/helpers/convertLegacyConfig.ts
