+++ packages/config/src/configProcessor/helpers/schemaManifest.ts
