+++ packages/config/src/configProcessor/index.ts
@@ -1,0 +1,5 @@
+import { convertJsonConfigToObject } from './converterJsonConfig';
+import { processConfig } from './processConfig';
+import { validateConfig } from './validateConfig';
+
+export { convertJsonConfigToObject, processConfig, validateConfig };
